Imgui Node Editor integration

+++ SynapseEngine/Engine/Animation/Processor/Geometry/AnimationColliderProcessor.cpp
@@ -136,8 +136,8 @@ namespace Syn
 
                     for (size_t l = 0; l < lodCount; ++l)
                     {
-                        uint32_t lodDescIndex = (static_cast<uint32_t>(m) * 4) + l;
-                        if (lodDescIndex >= model.meshletDrawDescriptors.size()) continue;
+                        uint32_t lodDescIndex = static_cast<uint32_t>((static_cast<uint32_t>(m) * 4) + l);
+                        if (lodDescIndex >= static_cast<uint32_t>(model.meshletDrawDescriptors.size())) continue;
 
                         const auto& drawDesc = model.meshletDrawDescriptors[lodDescIndex];
                         uint32_t meshletOffset = drawDesc.meshletOffset;
@@ -148,7 +148,7 @@ namespace Syn
 
                         for (size_t ml = 0; ml < meshletCount; ++ml)
                         {
-                            uint32_t globalMeshletIdx = meshletOffset + ml;
+                            uint32_t globalMeshletIdx = static_cast<uint32_t>(meshletOffset + ml);
                             const auto& meshletDesc = meshletDescs[globalMeshletIdx];
                             CookedAnimationFrameMeshlet& frameMeshlet = frameLod.meshlets[ml];
